+++ cirq/protocols/decompose_protocol.py
