+++ cirq/protocols/decompose_protocol_test.py
