+++ cirq/protocols/measurement_key_protocol.py
