+++ cirq/protocols/measurement_key_protocol_test.py
